test(canon): move the tsconfig extends cases into their own module
